+++ src/main/java/com/groupten/bmsproject/ProductionSchedule/ProductionScheduleEntity.java
@@ -11,8 +11,10 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.SequenceGenerator;
+import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.IntegerProperty;
 import javafx.beans.property.ObjectProperty;
+import javafx.beans.property.SimpleDoubleProperty;
 import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleObjectProperty;
 import javafx.beans.property.SimpleStringProperty;
@@ -26,7 +28,7 @@ public class ProductionScheduleEntity {
 
     private Integer id;
     private String productname;
-    private Integer quantity;
+    private Double quantity;
     private String lvlofstock;
     private LocalDate dateofproduction;
     private LocalDate expdate;
@@ -57,11 +59,11 @@ public class ProductionScheduleEntity {
         this.productname = productname;
     }
 
-    public Integer getproductschedQuantity() {
+    public Double getproductschedQuantity() {
         return quantity;
     }
 
-    public void setproductschedQuantity(Integer quantity) {
+    public void setproductschedQuantity(Double quantity) {
         this.quantity = quantity;
     }
 
@@ -110,8 +112,8 @@ public class ProductionScheduleEntity {
         return new SimpleStringProperty(productname);
     }
 
-    public IntegerProperty quantityProperty() {
-        return new SimpleIntegerProperty(quantity);
+    public DoubleProperty quantityProperty() {
+        return new SimpleDoubleProperty(quantity);
     }
 
     public StringProperty lvlofstockProperty() {
